Added barrelling for volume controls
